I really hate this fix, but there it is
For no good reason the UICollectionView is refusing
to use the correct UICollectionViewCell class after
calling 'reloadData'.  Unless I delay the reload by
0.3+ secs.  I tried 0.01, 0.1, 0.2, and the bug was
still there. 0.3 worked, but I didn't want to be so
close to the magic number, so I opted for 0.5s. oh,
and I want it to be noticed that every line of this
commit has the same number of chars, except this one